move steps to the rendering layout
