+++ libs/newsletter-workflow/src/lib/steps/renderingOptions/podcastLayout.ts
@@ -2,16 +2,20 @@ import type { WizardStepLayout } from '@newsletters-nx/state-machine';
 import { executeModify } from '../../executeModify';
 import { getStringValuesFromRecord } from '../../getValuesFromRecord';
 import { regExPatterns } from '../../regExPatterns';
-import { formSchemas } from './formSchemas';
+import { formSchemas } from '../newsletterData/formSchemas';
 
 const markdownTemplate = `
-# Specify the footer setup for {{name}}
+# Rendering Podcast sections in {{name}}
 
-What email address would you like to display in the footer of **{{name}}**?
+You may want to display some sections in **{{name}}** as Podcast sections.
 
-For example: newsletters@theguardian.com
+This will display the section with a dark background, and append a podcast graphic.
 
-![Email footer](https://i.guim.co.uk/img/uploads/2023/03/15/Email_address_contact.png?quality=85&dpr=2&width=300&s=98ce6a07791b0bfb45f2df9c732ea1d2)
+![Podcast styling](https://i.guim.co.uk/img/uploads/2023/03/15/podcast.png?quality=85&dpr=2&width=300&s=441a36a662b2adc04b867e3d617d7315)
+
+In order to use this styling for one or more sections, you need to specify the subheading for each section.
+
+The styling will be implemented wherever these subheadings are used.
 
 `.trim();
 
@@ -20,9 +24,9 @@ const staticMarkdown = markdownTemplate.replace(
 	'the newsletter',
 );
 
-export const footerLayout: WizardStepLayout = {
+export const podcastLayout: WizardStepLayout = {
 	staticMarkdown,
-	label: 'Footer Setup',
+	label: 'Podcast Sections',
 	dynamicMarkdown(requestData, responseData) {
 		if (!responseData) {
 			return staticMarkdown;
@@ -34,21 +38,15 @@ export const footerLayout: WizardStepLayout = {
 		back: {
 			buttonType: 'RED',
 			label: 'Back',
-			stepToMoveTo: 'podcast',
+			stepToMoveTo: 'linkList',
 			executeStep: executeModify,
 		},
 		finish: {
 			buttonType: 'GREEN',
 			label: 'Next',
-			stepToMoveTo: 'identityName',
-			onBeforeStepChangeValidate: (stepData): string | undefined => {
-				const email = stepData.formData
-					? stepData.formData['renderingOptions.contactEmail']
-					: undefined;
-				return email ? undefined : 'NO EMAIL ADDRESS PROVIDED';
-			},
+			stepToMoveTo: 'footer',
 			executeStep: executeModify,
 		},
 	},
-	schema: formSchemas.footer,
+	schema: formSchemas.podcast,
 };
